V1.1 - Shell
- Kernel will probably be untouched from now as few rough edges have
been patched.
- A basic shell is done with a couple of commands that demonstrates some
features of the kernel.

+++ kernel/inc/task.h
@@ -4,7 +4,7 @@
 
 typedef struct{
     uint32_t *stack_pointer;
-    volatile uint8_t delay_timer;
+    volatile uint32_t delay_timer;
     void (*task_function)(void);
 }task_t;
